Struct tnc_analysis_lib::prelude::Matrix

pub struct Matrix<T> {
    pub inner: T,
}

Main struct. Hosts the data required to move through the different phases of the TNC analysis.

Fields
inner: T
Implementations

impl Matrix<DataFrame>

fn new(inner: DataFrame) -> Self

pub fn to_row_dominant(&self, columns: &Vec<String>) -> Result<(Vec<f64>, usize)>

Converts the predictors in the matrix to a 1D array. The dummy variables need to be built here to avoid having to reference them post instantiation.

Dependency: User knows the original column names, pre-dummie making.

pub fn from_file(
path: &str,
header: Option<Header<'_>>
) -> Result<Matrix<DataFrame>>

pub fn with_propensity(&mut self, cfg: PropensityCfg) -> Result<()>

Appends a propensity field to the Matrix. Requires a configuration.

pub struct PropensityCfg<'a> {
    pub target: BinaryTarget<'a>,
    pub predictors: Predictors<'a>,
    pub bin_count: Option<u32>,
    pub name: Option<&'a str>,
}

pub fn write_to_file_csv<P: AsRef<Path>>(&mut self, path: P) -> Result<()>

pub fn header(&self) -> Header<'_>

pub fn with_include_tag(&mut self) -> (usize, usize)

Make this part of the initialization sequence

pub fn predictors(&self, cfg: FieldNamesCfg) -> Predictors<'_>

predictors Use “q_” & “derived” in the Matrix. Override with cfg. Used to specify the Logit model

pub fn binary_target(&self, cfg: FieldNamesCfg) -> BinaryTarget<'_>

pub fn write_to_file<P: AsRef<Path>>(&mut self, path: Option<P>) -> Result<()>

write so that target-binary is the first arrow

pub fn show_meta(&self) -> Result<String>

pub fn show_fields(&self) -> Result<String>